[ADD] lib.rs to manage logic of app
